Move database modules into subpackage
