+++ project-metrics/metrics_service/database/init_db.py
@@ -5,12 +5,12 @@ from absl import app
 from absl import flags
 import sqlalchemy
 
-import db_engine
+from database import db_engine
 # This import is required to ensure that subclasses of sqlalchemy.Base register
 # themselves in its metadata. This is how the create_all and drop_all methods
 # can identify which tables need to be created/dropped based on the defined ORM
 # models.
-import models  # pylint: disable=unused-import
+from database import models  # pylint: disable=unused-import
 
 
 FLAGS = flags.FLAGS
